Move Pipeline execution logic to _Node
For better coupling of helper functions
